+++ lib/screen/Quotation/enqScheduleQuotation.dart
@@ -59,6 +59,7 @@ class _EnqScheduleQuotationState extends State<EnqScheduleQuotation> {
     return Scaffold(
       floatingActionButton: Consumer<QuotationController>(
         builder: (context, value, child) => FloatingActionButton.extended(
+          backgroundColor: Colors.green,
           onPressed: () {
             Provider.of<ProductController>(context, listen: false)
                 .geProductList(context);
@@ -78,7 +79,9 @@ class _EnqScheduleQuotationState extends State<EnqScheduleQuotation> {
                         owner: value.c_person.toString(),
                         area: value.area.toString(),
                         enqId: value.enq_id.toString(),
-                        page: "enquiry schedule",type: "edit enq",rwId: "0",
+                        page: "enquiry schedule",
+                        type: "edit enq",
+                        rwId: "0",
                       )),
             );
           },
@@ -679,25 +682,31 @@ class _EnqScheduleQuotationState extends State<EnqScheduleQuotation> {
                                               children: [
                                                 InkWell(
                                                   onTap: () {
-                                                    DeletePopup popup =
-                                                        DeletePopup();
-                                                    popup.builddeletePopupDialog(
-                                                        context,
-                                                        value
-                                                            .quotProdItem[index]
-                                                                ["product_name"]
-                                                            .toString(),
-                                                        value
-                                                            .quotProdItem[index]
-                                                                ["product_id"]
-                                                            .toString(),
-                                                        index,
-                                                        "enqpdt",
-                                                        value.quotProdItem[
-                                                            index]["enq_id"],
-                                                        "",
-                                                        "",
-                                                        "","1","");
+                                                    if (value.quotProdItem
+                                                            .length >
+                                                        1) {
+                                                      DeletePopup popup =
+                                                          DeletePopup();
+                                                      popup.builddeletePopupDialog(
+                                                          context,
+                                                          value.quotProdItem[
+                                                                  index][
+                                                                  "product_name"]
+                                                              .toString(),
+                                                          value.quotProdItem[
+                                                                  index]
+                                                                  ["product_id"]
+                                                              .toString(),
+                                                          index,
+                                                          "enqpdt",
+                                                          value.quotProdItem[
+                                                              index]["enq_id"],
+                                                          "",
+                                                          "",
+                                                          "",
+                                                          "1",
+                                                          "");
+                                                    }
                                                   },
                                                   child: Row(
                                                     children: [
